requirements.txt
httpx
Pillow
python-telegram-bot